Add loading older messages when scrolling, messages load under socket
Create new socket event to load messages whet it is necessary. So, whem
you scroll to much and when you have just join the room, messages have
been loaded. Reorganize js file a bit to support changes. Add var to
config which declares the number of messages to load during the one
event